chore: 대대적 구조 개편
