--- src/main/java/com/example/demo/Book/BookRepository.java
@@ -1,14 +1,0 @@
-package com.example.demo.Book;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-import java.util.List;
-import java.util.Optional;
-
-@Repository
-public interface BookRepository extends JpaRepository<Book, Long> {
-    Book findById(long id);
-    Book findByTitleAndAuthorAndPublisher(String title, String author, String publisher);
-    List<Book> findByTitleContains(String title);
-}
