--- src/main/java/com/example/demo/Book/BookRequestForDifficulty.java
@@ -1,15 +1,0 @@
-package com.example.demo.Book;
-
-import lombok.Getter;
-import lombok.Setter;
-
-@Getter
-@Setter
-public class BookRequestForDifficulty {
-
-    private String difficulty;
-
-    public BookRequestForDifficulty(String difficulty) {
-        this.difficulty = difficulty;
-    }
-}
