--- src/main/java/com/example/demo/Book/BookResponseWithDifficulty.java
@@ -1,36 +1,0 @@
-package com.example.demo.Book;
-
-import lombok.Builder;
-import lombok.Getter;
-import lombok.Setter;
-
-import java.net.URL;
-
-@Builder
-@Getter
-@Setter
-public class BookResponseWithDifficulty {
-
-    private Long bookId;
-    private String title;
-    private String author;
-    private String publisher;
-    private int pageCount;
-    private URL thumbnail;
-    private float difficultyScore;
-    private String difficultyState;
-
-    public static BookResponseWithDifficulty from(BookDto bookDto) {
-        return BookResponseWithDifficulty.builder()
-                .bookId(bookDto.getId())
-                .title(bookDto.getTitle())
-                .author(bookDto.getAuthor())
-                .publisher(bookDto.getPublisher())
-                .pageCount(bookDto.getPageCount())
-                .thumbnail(bookDto.getThumbnail())
-                .difficultyScore(bookDto.getDifficultyScore())
-                .difficultyState(bookDto.getDifficultyState())
-                .build();
-    }
-
-}
