+++ src/main/java/com/example/demo/Book/Presentation/Request/BookRequestForDifficulty.java
@@ -1,0 +1,15 @@
+package com.example.demo.Book.Presentation.Request;
+
+import lombok.Getter;
+import lombok.Setter;
+
+@Getter
+@Setter
+public class BookRequestForDifficulty {
+
+    private String difficulty;
+
+    public BookRequestForDifficulty(String difficulty) {
+        this.difficulty = difficulty;
+    }
+}
