+++ src/main/java/com/example/demo/Book/Presentation/Request/BookRequestForListUp.java
@@ -1,0 +1,15 @@
+package com.example.demo.Book.Presentation.Request;
+
+import lombok.Getter;
+import lombok.Setter;
+
+@Getter
+@Setter
+public class BookRequestForListUp {
+    private String keyword;
+
+    public BookRequestForListUp(String keyword) {
+        this.keyword = keyword;
+    }
+
+}
