+++ src/main/java/com/example/demo/Book/Presentation/Response/BookResponseWithDifficulty.java
@@ -1,0 +1,37 @@
+package com.example.demo.Book.Presentation.Response;
+
+import com.example.demo.Book.Application.BookDto;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.Setter;
+
+import java.net.URL;
+
+@Builder
+@Getter
+@Setter
+public class BookResponseWithDifficulty {
+
+    private Long bookId;
+    private String title;
+    private String author;
+    private String publisher;
+    private int pageCount;
+    private URL thumbnail;
+    private float difficultyScore;
+    private String difficultyState;
+
+    public static BookResponseWithDifficulty from(BookDto bookDto) {
+        return BookResponseWithDifficulty.builder()
+                .bookId(bookDto.getId())
+                .title(bookDto.getTitle())
+                .author(bookDto.getAuthor())
+                .publisher(bookDto.getPublisher())
+                .pageCount(bookDto.getPageCount())
+                .thumbnail(bookDto.getThumbnail())
+                .difficultyScore(bookDto.getDifficultyScore())
+                .difficultyState(bookDto.getDifficultyState())
+                .build();
+    }
+
+}
