--- src/main/java/com/example/demo/BookMark/BookMarkRequestForTime.java
@@ -1,14 +1,0 @@
-package com.example.demo.BookMark;
-
-import lombok.Getter;
-import lombok.Setter;
-
-@Getter
-@Setter
-public class BookMarkRequestForTime {
-    private int time;
-
-    public BookMarkRequestForTime(int time) {
-        this.time = time;
-    }
-}
