+++ src/main/java/com/example/demo/BookMark/Presentation/Request/BookMarkRequestForTime.java
@@ -1,0 +1,14 @@
+package com.example.demo.BookMark.Presentation.Request;
+
+import lombok.Getter;
+import lombok.Setter;
+
+@Getter
+@Setter
+public class BookMarkRequestForTime {
+    private int time;
+
+    public BookMarkRequestForTime(int time) {
+        this.time = time;
+    }
+}
