+++ src/main/java/com/example/demo/User/Domain/User.java
@@ -1,0 +1,37 @@
+package com.example.demo.User.Domain;
+
+import com.example.demo.BookMark.Domain.BookMark;
+import com.example.demo.memo.Domain.Memo;
+import jakarta.persistence.*;
+import lombok.*;
+
+import java.util.ArrayList;
+import java.util.List;
+
+@Entity
+@Setter
+@Getter
+@Builder
+@AllArgsConstructor
+@NoArgsConstructor
+public class User {
+    @Id
+    private String email;
+    private String nickname;
+
+    @OneToMany(
+            mappedBy = "user",
+            fetch = FetchType.LAZY,
+            cascade = CascadeType.ALL,
+            orphanRemoval = true
+    )
+    private List<BookMark> bookmarks = new ArrayList<>();
+
+    @OneToMany(
+            mappedBy = "user",
+            fetch = FetchType.LAZY,
+            cascade = CascadeType.ALL,
+            orphanRemoval = true
+    )
+    private List<Memo> memos = new ArrayList<>();
+}
