+++ src/main/java/com/example/demo/User/Domain/UserRepository.java
@@ -1,0 +1,9 @@
+package com.example.demo.User.Domain;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface UserRepository extends JpaRepository<User, String> {
+    User findByEmail(String email);
+}
