--- src/main/java/com/example/demo/User/UserRepository.java
@@ -1,11 +1,0 @@
-package com.example.demo.User;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-import java.util.Optional;
-
-@Repository
-public interface UserRepository extends JpaRepository<User, String> {
-    User findByEmail(String email);
-}
