--- src/main/java/com/example/demo/memo/MemoRepository.java
@@ -1,14 +1,0 @@
-package com.example.demo.memo;
-
-import com.example.demo.Book.Book;
-import com.example.demo.User.User;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-import java.util.List;
-import java.util.Optional;
-
-@Repository
-public interface MemoRepository extends JpaRepository<Memo, Long> {
-    List<Memo> findByBookAndUser(Book book, User user);
-}
