--- src/main/java/com/example/demo/memo/MemoRequest.java
@@ -1,15 +1,0 @@
-package com.example.demo.memo;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.Setter;
-
-import java.net.URL;
-
-@Getter
-@Setter
-@AllArgsConstructor
-public class MemoRequest {
-    private String content;
-    private String date;
-}
