+++ src/main/java/com/example/demo/memo/Presentation/Request/MemoRequest.java
@@ -1,0 +1,13 @@
+package com.example.demo.memo.Presentation.Request;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.Setter;
+
+@Getter
+@Setter
+@AllArgsConstructor
+public class MemoRequest {
+    private String content;
+    private String date;
+}
