+++ src/main/java/com/example/demo/memo/Presentation/Response/MemoResponseAboutMemoId.java
@@ -1,0 +1,21 @@
+package com.example.demo.memo.Presentation.Response;
+
+import lombok.Builder;
+import lombok.Getter;
+import lombok.Setter;
+
+@Builder
+@Getter
+@Setter
+public class MemoResponseAboutMemoId {
+    private Long id;
+    private String message;
+
+    public static MemoResponseAboutMemoId from(Long id, String successMessage) {
+        return MemoResponseAboutMemoId.builder()
+                .id(id)
+                .message(successMessage)
+                .build();
+    }
+
+}
